refactor: split onboard write recovery
